updated orbit

+++ R/0D/Camera.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <glm/glm.hpp>
-
+#include <iostream>
 
 class Camera
 {
@@ -14,8 +14,8 @@ class Camera
   Camera(glm::vec3 p, glm::mat3 r, float f);
   void updateRotation(float X, float Y, float Z);
   void lookat(glm::vec3 point);
-  void forward(int amount);
-  void right(int amount);
-  void up(int amount);
+  void forward(float amount);
+  void right(float amount);
+  void up(float amount);
 
 };
